Demonstration of problem deep includes.
Simple project which simply invokes "std::min". However, on
Windows, through a series of nested includes, it manages to
include "Winsock2.h" (as winsock2.h, lowercase) prior to the
definition of "NOMINMAX".

This allows Winsock2.h to leak in the old MS macros for min
and max, causing 'std::min' in nesting.cpp to break.

MSVC already provides a way to track this problem down, by
turning on "Show Includes" (/showIncludes), but in a very large
project this can still take a lot of time to track down.

In this project, it's trivial to locate the problem because
there is only one source file and only one project. In a large
scale project it becomes much harder because MSVC does not
provide an easy way to correlate an include error with a
specific source file.

Additionally, you cannot click on a filename in the output window
and get to that file, which is a small bug irritating inconvenience
when dealing with an issue like this.